# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Game Editor API
  description: This API describes representing objects with GeoJson from PostGIS database.
# during dev, should point to your local machine
host: game.iamborsch.ru
# basePath prefixes all resource paths 
basePath: /api/v1
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /points:
    x-swagger-router-controller: points
    get:
      description: Get feature by point coordinates.
      operationId: getFeatureByPoint
      parameters:
        - name: latitude
          in: query
          description: Latitude component of point.
          required: true
          type: integer
        - name: longitude
          in: query
          description: Longitude component of point.
          required: true
          type: integer
        - name: altitude
          in: query
          description: Altitude component of point.
          required: true
          type: integer
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Points
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Feature"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      description: Add new feature object.
      operationId: createFeature
      parameters:
        - name: feature
          in: body
          schema:
            $ref: "#/definitions/Feature"
          required: true
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Points
      responses:
        201:
          description: OK
          schema:
            $ref: "#/definitions/Feature"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /points/{id}/:
    x-swagger-router-controller: points
    get:
      description: Get feature by id.
      operationId: getFeatureById
      parameters:
        - name: id
          type: integer
          in: path
          required: true
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Points
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Feature"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    put:
      description: Update feature by id.
      operationId: updateFeature
      parameters:
        - name: id
          type: integer
          in: path
          required: true
        - name: feature
          in: body
          schema:
            $ref: "#/definitions/Feature"
          required: true
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Points
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Feature"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      description: Delete feature by id
      operationId: delFeature
      tags:
        - Points
      parameters:
        - name: id
          type: integer
          in: path
          required: true
      responses:
        200:
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    
  /locations:
    x-swagger-router-controller: locations
    get:
      description: Returns all available features in box area or returns all features if box parameters not defined (test mode).
      operationId: getAll
      parameters:
        - name: b_top
          in: query
          description: Top point of box.
          type: integer
        - name: b_left
          in: query
          description: Left point of box.
          type: integer
        - name: b_right
          in: query
          description: Right point of box.
          type: integer
        - name: b_bottom
          in: query
          description: Bottom point of box.
          type: integer
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Locations
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FeatureCollection"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  FeatureCollection:
    type: object
    required:
      - type
      - features
    description: This is a collection of features. [GeoJson FeatureCollection] (http://geojson.org/geojson-spec.html#examples)
    properties:
      type:
        type: string
        default: FeatureCollection
      features:
        type: array
        items:
          $ref: "#/definitions/Feature"
  Feature:
    type: object
    required:
      - type
      - geometry
      - properties
    description: Elementary objects with geometry properties.
    properties:
      type:
        type: string
        default: Feature
      bbox:
        type: array
        items: {}
        description: http://geojson.org/geojson-spec.html#bounding-boxes
      geometry:
        $ref: "#/definitions/Geometry"
      properties:
        type: object
        properties:
          id:
            type: integer
            description: this is a geometry id in database.
          name:
            type: string
            description: name of this geometry in database.
  Geometry:
    type: object
    required:
      - type
      - coordinates
    properties:
      type:
        type: string
        default: Point
        enum: [
          Point,
          LineString,
          Polygon,
          Multipoint,
          MultiLineString,
          MultiPolygon,
          GeometryCollection
        ]
        description: "[GeoJSON geometries] (http://geojson.org/geojson-spec.html#geometry-objects)"
      coordinates:
        type: array
        items: {}
        description: x, y, z coordinates in array of geometry object. I can't describe all possible variations, so look at GeoJson docs.
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string